Stub NewDocMethods
